+++ History.md
@@ -1172,3 +1172,17 @@ commit by afterloe (lm6289511@gmail.com)
 > services/userService.js  
 > tools/lib/util.js  
 
+
+1.0.3 / 2017-2-7 16:56:38
+==================
+commit by afterloe (lm6289511@gmail.com)
+
+  * master: ws前端接口修改，内部样式提升，添加了额工具中部分方法
+
+
+### modify files
+> servers/websocket.js  
+> services/userService.js  
+> tools/_util.html  
+> tools/lib/util.js  
+